requirements.txt
discord.py==2.7.3
PyNaCl==1.5.0
yt-dlp==2025.12.21
python-dotenv==1.0.1
